requirements.txt
matplotlib==3.9.0
numpy==2.0.0
python-dotenv==1.0.1
requests==2.32.3
pytest==8.2.2
pytest-mock==3.14.0
black==24.4.2
flake8==7.1.0